Refactor channel diffusion state updates
