+++ src/heart/renderers/channel_diffusion/__init__.py
@@ -1,3 +1,5 @@
+from heart.renderers.channel_diffusion.provider import \
+    ChannelDiffusionStateProvider  # noqa: F401
 from heart.renderers.channel_diffusion.renderer import \
     ChannelDiffusionRenderer  # noqa: F401
 from heart.renderers.channel_diffusion.state import \
