+++ src/heart/renderers/channel_diffusion/renderer.py
@@ -1,20 +1,22 @@
 from __future__ import annotations
 
-import numpy as np
 import pygame
 
 from heart import DeviceDisplayMode
 from heart.device import Orientation
 from heart.peripheral.core.manager import PeripheralManager
 from heart.renderers import AtomicBaseRenderer
+from heart.renderers.channel_diffusion.provider import \
+    ChannelDiffusionStateProvider
 from heart.renderers.channel_diffusion.state import ChannelDiffusionState
 
 
 class ChannelDiffusionRenderer(AtomicBaseRenderer[ChannelDiffusionState]):
-    def __init__(self) -> None:
+    def __init__(self, provider: ChannelDiffusionStateProvider | None = None) -> None:
         super().__init__()
         self.device_display_mode = DeviceDisplayMode.FULL
         self.warmup = False
+        self._provider = provider or ChannelDiffusionStateProvider()
 
     def _create_initial_state(
         self,
@@ -24,14 +26,7 @@ class ChannelDiffusionRenderer(AtomicBaseRenderer[ChannelDiffusionState]):
         orientation: Orientation,
     ) -> ChannelDiffusionState:
         width, height = window.get_size()
-        grid = np.zeros((width, height, 3), dtype=np.uint8)
-        grid[width // 2, height // 2] = np.array([255, 255, 255], dtype=np.uint8)
-        return ChannelDiffusionState(grid=grid)
-
-    def _compute_center_after_fade(self, grid: np.ndarray) -> np.ndarray:
-        brightness = grid.max(axis=2)
-        faded = grid - (brightness // 2)[:, :, None]
-        return faded
+        return self._provider.initial_state(width=width, height=height)
 
     def real_process(
         self,
@@ -39,29 +34,7 @@ class ChannelDiffusionRenderer(AtomicBaseRenderer[ChannelDiffusionState]):
         clock: pygame.time.Clock,
         orientation: Orientation,
     ) -> None:
-        grid = self.state.grid.astype(np.int32)
+        next_state = self._provider.next_state(self.state)
+        self.set_state(next_state)
 
-        red = grid[:, :, 0]
-        green = grid[:, :, 1]
-        blue = grid[:, :, 2]
-
-        new_grid = np.zeros_like(grid)
-
-        center = self._compute_center_after_fade(grid)
-        new_grid += center
-
-        new_grid[:, :-1, 1] += green[:, 1:]
-        new_grid[:, 1:, 1] += green[:, :-1]
-
-        new_grid[:-1, :, 2] += blue[1:, :]
-        new_grid[1:, :, 2] += blue[:-1, :]
-
-        new_grid[:-1, :-1, 0] += red[1:, 1:]
-        new_grid[1:, :-1, 0] += red[:-1, 1:]
-        new_grid[:-1, 1:, 0] += red[1:, :-1]
-        new_grid[1:, 1:, 0] += red[:-1, :-1]
-
-        clipped_grid = np.clip(new_grid, 0, 255).astype(np.uint8)
-        self.set_state(ChannelDiffusionState(grid=clipped_grid))
-
-        pygame.surfarray.blit_array(window, clipped_grid)
+        pygame.surfarray.blit_array(window, next_state.grid)
